+++ modules/system_manager.mjs
@@ -4,15 +4,18 @@ import {
   AnimManager,
   Star,
   Planet,
+  EllipticalTrajectory
 } from '../exo3d.mjs'
 
 class GravitationalSystemManager {
   light
   skybox
   constructor(scene, UI, canvas) {
-    const V_ORIGIN = new BABYLON.Vector3(0, 0, 0) // Origine du repère
-    const TRAJECTORY_RADIUS = 8
-    const PRECISION_STEPS = 100
+    const PRECISION_STEPS = 300 // Number of points for the static trajectory
+    const a = 8 // Semi major axis, because it is a very important parameter, I choose to give it its original name "a"
+    const e = 0 // Excentricty, can be anything between 0 included and 1 excluded
+    const b = Math.sqrt(Math.pow(a, 2) * (1 - Math.pow(e, 2))) // Value of b based on a and e
+    const V_ORIGIN_STAR = new BABYLON.Vector3(-2 * a * e, 0, 0) // Position of the star, also it is the left focus of the ellipse
 
     // Textures' source : https://www.solarsystemscope.com/textures/
     // BUG : earth texture is upside down for some reason, need to be fixed but it can wait the fake svg textures fix
@@ -27,13 +30,13 @@ class GravitationalSystemManager {
       texture: STAR_TEXTURE,
       distanceToParent: 0,
       color: starColor,
-      originalPosition: V_ORIGIN,
-      inclinationAngle: 0, // inclination and temperature aren't important now, but they're ready for the next features
-      temperature: 5000,
+      originalPosition: V_ORIGIN_STAR,
+      inclinationAngle: 0, // Inclination and temperature aren't important now, but they're ready for the next features
+      temperature: 5000
     }
 
     const planetColor = new BABYLON.Color3(0.5, 0.3, 0.3) // Arbitrary color (brown), not in caps because it will depend on other parameters
-    const PLANET_POSITION = new BABYLON.Vector3(TRAJECTORY_RADIUS, 0, 0)
+    const PLANET_POSITION = new BABYLON.Vector3(a, 0, 0)
     const planetOptions = {
       diameter: 1,
       texture: PLANET_TEXTURE,
@@ -41,22 +44,22 @@ class GravitationalSystemManager {
       color: planetColor,
       originalPosition: PLANET_POSITION,
       inclinationAngle: 0,
-      temperature: 0, // may seem cold, but it's not as cold as my office right now
+      temperature: 0 // May seem cold, but it's not as cold as my office right now
     }
 
     const satelliteOptions = {
       diameter: 0.3,
       texture: SATELLITE_TEXTURE,
       distanceToParent: 2, // Arbitrary position relative to the planet
-      color: undefined, // satellite does not need a color, only a texture
-      originalPosition: undefined, // same for its original position, as it will be modified just as it becomes a child of the planet
+      color: undefined, // Satellite does not need a color, only a texture
+      originalPosition: undefined, // Same for its original position, as it will be modified just as it becomes a child of the planet
       inclinationAngle: 0,
-      temperature: 0,
+      temperature: 0
     }
 
     const star = new Star(starOptions, scene)
     const planet = new Planet(planetOptions, scene)
-    const satellite = new Planet(satelliteOptions, scene) // satellite is considered a planet because it has the same parameters right now
+    const satellite = new Planet(satelliteOptions, scene) // Satellite is considered a planet because it has the same parameters right now
 
     satellite.mesh.parent = planet.mesh
     satellite.mesh.position = new BABYLON.Vector3(
@@ -65,27 +68,22 @@ class GravitationalSystemManager {
       0
     )
 
-    const trajectory = new Array(PRECISION_STEPS + 1) // Il faut compter un point supplémentaire pour fermer la trajectoire
-
-    // TODO : faire une fonction qui calcule n'importe quelle trajectoire elliptique, pas juste un cercle
-    for (const i of trajectory.keys()) {
-      trajectory[i] = new BABYLON.Vector3(
-        TRAJECTORY_RADIUS * Math.cos((Math.PI / (PRECISION_STEPS / 2)) * i),
-        0,
-        TRAJECTORY_RADIUS * Math.sin((Math.PI / (PRECISION_STEPS / 2)) * i)
-      )
-    }
-    let trajectoryLine = BABYLON.MeshBuilder.CreateLines('trajectory', {
-      points: trajectory,
+    const trajectory = new EllipticalTrajectory({
+      a: a,
+      b: b
     })
-    trajectoryLine.color = new BABYLON.Color4(1, 0, 0, 1)
+    const staticTrajectory = trajectory.staticTrajectory(PRECISION_STEPS)
 
-    // De nouveaux paramètres pourront être ajoutés pour les visuels, comme la température de l'étoile (pour sa couleur) ou la luminosité souhaitée
+    let trajectoryLine = BABYLON.MeshBuilder.CreateLines('trajectory', {
+      points: staticTrajectory
+    })
+    trajectoryLine.color = new BABYLON.Color3(1, 0, 0)
+
     const animationParameters = {
-      path: trajectory,
+      path: staticTrajectory,
       planet: planet,
       satellite: satellite,
-      steps: PRECISION_STEPS,
+      steps: PRECISION_STEPS
     }
 
     const anims = new Animations(scene, animationParameters)
@@ -97,9 +95,9 @@ class GravitationalSystemManager {
 
     // Partie lumière/brillance de l'étoile
 
-    this.light = new BABYLON.PointLight('light', V_ORIGIN)
+    this.light = new BABYLON.PointLight('light', V_ORIGIN_STAR)
     this.light.diffuse = star.color // Couleur projetée sur les objets autour de l'étoile
-    this.light.specular = new BABYLON.Color3.Black() //Empêche les reflets de type "boule de billard"
+    this.light.specular = new BABYLON.Color3.Black() // Empêche les reflets de type "boule de billard"
     this.light.range = 100 // Ce paramètre doit être soigneusement retenu, c'est ce qui permettra d'éclairer - ou pas - les objets éloignés de l'étoile
 
     const gl = new BABYLON.GlowLayer('glow', scene)